% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_functions.R
\name{fetch_smooth}
\alias{fetch_smooth}
\title{Calculate 'smoothed' wind fetch}
\usage{
fetch_smooth(raster, angle, n = 3, interval = 3)
}
\arguments{
\item{raster}{RasterLayer object.}

\item{angle}{RasterLayer object.}
}
\value{
raster RasterLayer or RasterStack object.
}
\description{
Function to calculate wind fetch across a surface. Wind fetch is the unobstructed travel length of wind across a surface from a given direction. Input surfaces are supplied as logical (TRUE/FALSE) RasterLayer objects from the raster package and angles as numeric scalars or vectors.
When only one angle is supplied a RasterLayer is returned and a RasterStack otherwise.
}
\details{
Smoothed wind fetch is the fetch a given angle as an average of fetch from 'n' angles spaced by 'interval' degrees which are added to each side of 'angle'.

Function uses future_lapply meaning that parallel processing can be enabled by calling future::plan(multisession) beforehand.
}
